🙈 worl/communities  and fuction getAllCommunities
